Refactoring MeshToFlow pass out from patterns. This moves the pass out to the Transforms/ directory and separates the pass logic from the patterns for consistency with all the other conversions. This will also help us move around where the pass is run and possibly merge it with the main ConvertToFlow pass by separating lookup from channel materialization as we do with device globals. Such refactorings will (likely) be required for proper multi-device scheduling but this initial refactoring is just moving code around.
