+++ compiler/src/iree/compiler/Dialect/Flow/Conversion/MeshToFlow/BUILD.bazel
@@ -4,7 +4,7 @@
 # See https://llvm.org/LICENSE.txt for license information.
 # SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-load("//build_tools/bazel:build_defs.oss.bzl", "iree_compiler_cc_library", "iree_gentbl_cc_library")
+load("//build_tools/bazel:build_defs.oss.bzl", "iree_compiler_cc_library")
 
 package(
     default_visibility = ["//visibility:public"],
@@ -15,13 +15,12 @@ package(
 iree_compiler_cc_library(
     name = "MeshToFlow",
     srcs = [
-        "MeshToFlow.cpp",
+        "Patterns.cpp",
     ],
     hdrs = [
-        "MeshToFlow.h",
+        "Patterns.h",
     ],
     deps = [
-        ":PassesIncGen",
         "//compiler/src/iree/compiler/Dialect/Flow/IR",
         "//compiler/src/iree/compiler/Dialect/Util/IR",
         "//compiler/src/iree/compiler/Utils",
@@ -34,25 +33,8 @@ iree_compiler_cc_library(
         "@llvm-project//mlir:LinalgUtils",
         "@llvm-project//mlir:MeshDialect",
         "@llvm-project//mlir:MeshTransforms",
-        "@llvm-project//mlir:Pass",
         "@llvm-project//mlir:Support",
         "@llvm-project//mlir:TensorDialect",
         "@llvm-project//mlir:TransformUtils",
-        "@llvm-project//mlir:Transforms",
-    ],
-)
-
-iree_gentbl_cc_library(
-    name = "PassesIncGen",
-    tbl_outs = [
-        (
-            ["--gen-pass-decls"],
-            "Passes.h.inc",
-        ),
-    ],
-    tblgen = "@llvm-project//mlir:mlir-tblgen",
-    td_file = "Passes.td",
-    deps = [
-        "@llvm-project//mlir:PassBaseTdFiles",
     ],
 )
