+++ compiler/src/iree/compiler/Dialect/Flow/Conversion/MeshToFlow/Patterns.cpp
@@ -4,40 +4,21 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "iree/compiler/Dialect/Flow/Conversion/MeshToFlow/MeshToFlow.h"
+#include "iree/compiler/Dialect/Flow/Conversion/MeshToFlow/Patterns.h"
 
 #include <algorithm>
-#include <cassert>
-#include <functional>
-#include <iterator>
 #include <numeric>
-#include <string>
-#include <tuple>
-#include <unordered_set>
-#include <utility>
 
 #include "iree/compiler/Dialect/Flow/IR/FlowDialect.h"
 #include "iree/compiler/Dialect/Flow/IR/FlowOps.h"
 #include "iree/compiler/Dialect/Flow/IR/FlowTypes.h"
 #include "iree/compiler/Dialect/Util/IR/UtilOps.h"
 #include "iree/compiler/Dialect/Util/IR/UtilTypes.h"
-#include "iree/compiler/Utils/Folding.h"
-#include "iree/compiler/Utils/Indexing.h"
-#include "iree/compiler/Utils/OpVisitor.h"
 #include "iree/compiler/Utils/Permutation.h"
-#include "iree/compiler/Utils/SmallVectorDenseMapInfo.h"
-#include "llvm/ADT/ArrayRef.h"
-#include "llvm/ADT/STLExtras.h"
-#include "llvm/ADT/SmallString.h"
-#include "llvm/ADT/SmallVector.h"
-#include "llvm/ADT/iterator_range.h"
-#include "llvm/Support/Casting.h"
 #include "llvm/Support/Debug.h"
-#include "llvm/Support/raw_ostream.h"
 #include "mlir/Dialect/Affine/IR/AffineOps.h"
 #include "mlir/Dialect/Arith/Utils/Utils.h"
 #include "mlir/Dialect/Linalg/IR/Linalg.h"
-#include "mlir/Dialect/Mesh/IR/MeshDialect.h"
 #include "mlir/Dialect/Mesh/IR/MeshOps.h"
 #include "mlir/Dialect/Mesh/Transforms/Simplifications.h"
 #include "mlir/Dialect/Mesh/Transforms/Transforms.h"
@@ -48,63 +29,12 @@
 #include "mlir/IR/BuiltinAttributes.h"
 #include "mlir/IR/BuiltinTypeInterfaces.h"
 #include "mlir/IR/BuiltinTypes.h"
-#include "mlir/IR/ImplicitLocOpBuilder.h"
 #include "mlir/IR/OpDefinition.h"
-#include "mlir/IR/PatternMatch.h"
 #include "mlir/IR/SymbolTable.h"
 #include "mlir/IR/Value.h"
-#include "mlir/IR/Visitors.h"
-#include "mlir/Pass/Pass.h"
-#include "mlir/Pass/PassManager.h"
-#include "mlir/Pass/PassRegistry.h"
 #include "mlir/Support/LLVM.h"
-#include "mlir/Support/LogicalResult.h"
-#include "mlir/Transforms/GreedyPatternRewriteDriver.h"
-
-#define DEBUG_TYPE "iree-mesh-to-flow"
 
 namespace mlir::iree_compiler::IREE::Flow {
-
-#define GEN_PASS_CLASSES
-#define GEN_PASS_REGISTRATION
-#include "iree/compiler/Dialect/Flow/Conversion/MeshToFlow/Passes.h.inc" // IWYU pragma: keep
-
-static bool hasMoreThanOneMesh(Operation *op) {
-  int meshCount = 0;
-  op->walk([&meshCount](mesh::MeshOp mesh) {
-    ++meshCount;
-    return meshCount > 1 ? WalkResult::interrupt() : WalkResult::advance();
-  });
-  return meshCount > 1;
-}
-
-static bool isDefaultChannel(mesh::MeshOp mesh,
-                             ArrayRef<mesh::MeshAxis> meshAxes) {
-  if (mesh.getRank() != static_cast<int64_t>(meshAxes.size())) {
-    return false;
-  }
-  return isIdentityPermutation(meshAxes);
-}
-
-static Value getDefaultChannel(mesh::MeshOp mesh, bool useNamedDefaultChannels,
-                               ImplicitLocOpBuilder &builder) {
-  if (useNamedDefaultChannels)
-    return builder.create<IREE::Flow::ChannelDefaultOp>(mesh.getSymName());
-  else
-    return builder.create<IREE::Flow::ChannelDefaultOp>();
-}
-
-// Remove from `values` elements that have indices present in filter.
-static SmallVector<Value> filterOutByIndex(ArrayRef<Value> values,
-                                           ArrayRef<mesh::MeshAxis> filter) {
-  SmallVector<Value> res;
-  for (size_t i = 0; i < values.size(); ++i) {
-    if (!llvm::is_contained(filter, i)) {
-      res.push_back(values[i]);
-    }
-  }
-  return res;
-}
 
 static CollectiveReductionOp
 convertReductionKind(mesh::ReductionKind reduction) {
@@ -125,132 +55,6 @@ static CollectiveReductionOpAttr
 convertReductionKind(mesh::ReductionKindAttr reduction) {
   return CollectiveReductionOpAttr::get(
       reduction.getContext(), convertReductionKind(reduction.getValue()));
-}
-
-static Value buildChannelCreation(mesh::MeshOp mesh,
-                                  ArrayRef<mesh::MeshAxis> meshAxes,
-                                  bool useNamedDefaultChannels,
-                                  ImplicitLocOpBuilder &builder) {
-  assert(mesh);
-  Value meshChannel = getDefaultChannel(mesh, useNamedDefaultChannels, builder);
-  SmallVector<Value> meshProcessMultiIndex =
-      builder.create<mesh::ProcessMultiIndexOp>(mesh).getResults();
-  SmallVector<Value> meshShape =
-      builder.create<mesh::MeshShapeOp>(mesh).getResults();
-  SmallVector<Value> reorderedMeshIndex =
-      permute(ArrayRef<Value>(meshProcessMultiIndex), meshAxes);
-  SmallVector<Value> reorderedMeshShape =
-      permute(ArrayRef<Value>(meshShape), meshAxes);
-  SmallVector<Value> groupIndex =
-      filterOutByIndex(meshProcessMultiIndex, meshAxes);
-  SmallVector<Value> groupsShape = filterOutByIndex(meshShape, meshAxes);
-  OpFoldResult reorderedProcessLinearIndex =
-      linearIndexFromShape(toOpFoldResults(reorderedMeshIndex),
-                           toOpFoldResults(reorderedMeshShape), builder);
-  OpFoldResult color = linearIndexFromShape(
-      toOpFoldResults(groupIndex), toOpFoldResults(groupsShape), builder);
-  return builder.create<IREE::Flow::ChannelSplitOp>(
-      meshChannel,
-      getValueOrCreateConstantIndexOp(builder, builder.getLoc(), color),
-      getValueOrCreateConstantIndexOp(builder, builder.getLoc(),
-                                      reorderedProcessLinearIndex));
-}
-
-static SmallString<64> getChannelName(mesh::MeshOp mesh,
-                                      ArrayRef<mesh::MeshAxis> axes) {
-  SmallString<64> res;
-  llvm::raw_svector_ostream stream(res);
-  stream << "_mesh_" << mesh.getSymName();
-  if (axes.empty()) {
-    return res;
-  }
-
-  stream << "_axes";
-  for (mesh::MeshAxis axis : axes) {
-    stream << "_" << axis;
-  }
-
-  return res;
-}
-
-static void buildChannelInitializer(mesh::MeshOp mesh,
-                                    ArrayRef<mesh::MeshAxis> meshAxes,
-                                    bool useNamedDefaultChannels,
-                                    ImplicitLocOpBuilder &builder) {
-  IREE::Util::InitializerOp initOp =
-      builder.create<IREE::Util::InitializerOp>();
-  Block *block = builder.createBlock(&initOp.getBody());
-  ImplicitLocOpBuilder::InsertionGuard insertionGuard(builder);
-  builder.setInsertionPointToStart(block);
-  Value channel =
-      buildChannelCreation(mesh, meshAxes, useNamedDefaultChannels, builder);
-  builder.create<IREE::Util::GlobalStoreOp>(channel,
-                                            getChannelName(mesh, meshAxes));
-  builder.create<IREE::Util::ReturnOp>();
-}
-
-// Construct a Flow channel inside `module` using
-// util.global and util.initializer.
-static void buildGlobalChannelCreation(mesh::MeshOp mesh,
-                                       ArrayRef<mesh::MeshAxis> meshAxes,
-                                       bool useNamedDefaultChannels,
-                                       ModuleOp module, OpBuilder &opBuilder) {
-  if (isDefaultChannel(mesh, meshAxes)) {
-    return;
-  }
-
-  ImplicitLocOpBuilder builder(mesh.getLoc(), opBuilder);
-  ImplicitLocOpBuilder::InsertionGuard insertionGuard(builder);
-  builder.setInsertionPointToStart(&module.getBodyRegion().getBlocks().front());
-
-  auto channelName = getChannelName(mesh, meshAxes);
-  builder.create<IREE::Util::GlobalOp>(
-      builder.getStringAttr("private"), channelName,
-      builder.getType<IREE::Flow::ChannelType>(), false, TypedAttr(),
-      builder.getAttr<IREE::Util::InlineNeverAttr>());
-  buildChannelInitializer(mesh, meshAxes, useNamedDefaultChannels, builder);
-}
-
-static Value buildCachedChannelLoading(mesh::MeshOp mesh,
-                                       ArrayRef<mesh::MeshAxis> meshAxes,
-                                       bool useNamedDefaultChannels,
-                                       ImplicitLocOpBuilder &builder) {
-  if (isDefaultChannel(mesh, meshAxes)) {
-    return getDefaultChannel(mesh, useNamedDefaultChannels, builder);
-  }
-  return builder.create<IREE::Util::GlobalLoadOp>(
-      IREE::Flow::ChannelType::get(builder.getContext()),
-      getChannelName(mesh, meshAxes));
-}
-
-// The !flow.channel corresponding to the mesh and mesh axes used in the op.
-template <typename MeshCollectiveOp>
-static Value buildCachedChannelLoading(
-    MeshCollectiveOp op, SymbolTableCollection &symbolTableCollection,
-    bool useNamedDefaultChannels, ImplicitLocOpBuilder &builder) {
-  ImplicitLocOpBuilder::InsertionGuard insertionGuard(builder);
-  builder.setInsertionPointAfter(op);
-
-  mesh::MeshOp mesh = mesh::getMesh(op, symbolTableCollection);
-  return buildCachedChannelLoading(mesh, op.getMeshAxes(),
-                                   useNamedDefaultChannels, builder);
-}
-
-SmallVector<mesh::MeshAxis> getAllMeshAxes(mesh::MeshOp mesh) {
-  SmallVector<mesh::MeshAxis> res(mesh.getRank());
-  std::iota(res.begin(), res.end(), 0);
-  return res;
-}
-
-static Value buildCachedChannelLoading(
-    mesh::ProcessLinearIndexOp op, SymbolTableCollection &symbolTableCollection,
-    bool useNamedDefaultChannels, ImplicitLocOpBuilder &builder) {
-  ImplicitLocOpBuilder::InsertionGuard insertionGuard(builder);
-  builder.setInsertionPointAfter(op);
-
-  mesh::MeshOp mesh = mesh::getMesh(op, symbolTableCollection);
-  return buildCachedChannelLoading(mesh, getAllMeshAxes(mesh),
-                                   useNamedDefaultChannels, builder);
 }
 
 static TypedValue<RankedTensorType>
@@ -384,29 +188,41 @@ struct MeshToFlowCollectiveRewritePatternBase : OpRewritePattern<Op> {
   template <typename... OpRewritePatternArgs>
   MeshToFlowCollectiveRewritePatternBase(
       SymbolTableCollection &symbolTableCollection,
-      bool useNamedDefaultChannels,
+      LookupMeshChannelFn lookupChannel,
       OpRewritePatternArgs &&...opRewritePatternArgs)
       : OpRewritePattern<Op>(
             std::forward<OpRewritePatternArgs...>(opRewritePatternArgs)...),
         symbolTableCollection(symbolTableCollection),
-        useNamedDefaultChannels(useNamedDefaultChannels) {}
+        lookupChannel(lookupChannel) {}
 
 protected:
+  // The !flow.channel corresponding to the mesh and mesh axes used in the op.
+  template <typename MeshCollectiveOp>
+  Value buildCachedChannelLoading(MeshCollectiveOp op,
+                                  ImplicitLocOpBuilder &builder) const {
+    mesh::MeshOp mesh = mesh::getMesh(op, symbolTableCollection);
+    return lookupChannel(builder.getLoc(), mesh, op.getMeshAxes(), builder);
+  }
+
+  Value buildCachedChannelLoading(mesh::ProcessLinearIndexOp op,
+                                  ImplicitLocOpBuilder &builder) const {
+    mesh::MeshOp mesh = mesh::getMesh(op, symbolTableCollection);
+    return lookupChannel(builder.getLoc(), mesh, std::nullopt, builder);
+  }
+
   SymbolTableCollection &symbolTableCollection;
-  bool useNamedDefaultChannels;
+  LookupMeshChannelFn lookupChannel;
 };
 
 struct MeshAllReduceToFlow
     : MeshToFlowCollectiveRewritePatternBase<mesh::AllReduceOp> {
   using MeshToFlowCollectiveRewritePatternBase::
       MeshToFlowCollectiveRewritePatternBase;
-
   LogicalResult matchAndRewrite(mesh::AllReduceOp op,
                                 PatternRewriter &rewriter) const override {
     ImplicitLocOpBuilder builder(op->getLoc(), rewriter);
     builder.setInsertionPointAfter(op.getOperation());
-    Value channel = buildCachedChannelLoading(op, symbolTableCollection,
-                                              useNamedDefaultChannels, builder);
+    Value channel = buildCachedChannelLoading(op, builder);
     Value target = builder.create<tensor::EmptyOp>(
         op.getResult().getType().getShape(),
         op.getResult().getType().getElementType());
@@ -424,12 +240,10 @@ struct MeshAllGatherToFlow
     : MeshToFlowCollectiveRewritePatternBase<mesh::AllGatherOp> {
   using MeshToFlowCollectiveRewritePatternBase::
       MeshToFlowCollectiveRewritePatternBase;
-
   LogicalResult matchAndRewrite(mesh::AllGatherOp op,
                                 PatternRewriter &rewriter) const override {
-    if (ShapedType::isDynamicShape(
-            cast<RankedTensorType>(op.getOperand().getType()).getShape()) ||
-        ShapedType::isDynamicShape(op.getResult().getType().getShape())) {
+    if (!cast<RankedTensorType>(op.getOperand().getType()).hasStaticShape() ||
+        !op.getResult().getType().hasStaticShape()) {
       // TODO: add dynamic support.
       return rewriter.notifyMatchFailure(op->getLoc(),
                                          "Dynamic tensor case is unsupported.");
@@ -437,8 +251,7 @@ struct MeshAllGatherToFlow
 
     ImplicitLocOpBuilder builder(op->getLoc(), rewriter);
     builder.setInsertionPointAfter(op.getOperation());
-    Value channel = buildCachedChannelLoading(op, symbolTableCollection,
-                                              useNamedDefaultChannels, builder);
+    Value channel = buildCachedChannelLoading(op, builder);
 
     int64_t gatherAxis = op.getGatherAxis().getSExtValue();
 
@@ -468,12 +281,10 @@ struct MeshAllToAllToFlow
     : MeshToFlowCollectiveRewritePatternBase<mesh::AllToAllOp> {
   using MeshToFlowCollectiveRewritePatternBase::
       MeshToFlowCollectiveRewritePatternBase;
-
   LogicalResult matchAndRewrite(mesh::AllToAllOp op,
                                 PatternRewriter &rewriter) const override {
-    if (ShapedType::isDynamicShape(
-            cast<RankedTensorType>(op.getOperand().getType()).getShape()) ||
-        ShapedType::isDynamicShape(op.getResult().getType().getShape())) {
+    if (!cast<RankedTensorType>(op.getOperand().getType()).hasStaticShape() ||
+        !op.getResult().getType().hasStaticShape()) {
       // TODO: add dynamic support.
       return rewriter.notifyMatchFailure(op->getLoc(),
                                          "Dynamic tensor case is unsupported.");
@@ -494,8 +305,7 @@ struct MeshAllToAllToFlow
     ImplicitLocOpBuilder builder(op->getLoc(), rewriter);
     builder.setInsertionPointAfter(op.getOperation());
 
-    Value channel = buildCachedChannelLoading(op, symbolTableCollection,
-                                              useNamedDefaultChannels, builder);
+    Value channel = buildCachedChannelLoading(op, builder);
 
     int64_t splitAxis = op.getSplitAxis().getSExtValue();
 
@@ -526,13 +336,11 @@ struct MeshProcessLinearIndexToFlow
     : MeshToFlowCollectiveRewritePatternBase<mesh::ProcessLinearIndexOp> {
   using MeshToFlowCollectiveRewritePatternBase::
       MeshToFlowCollectiveRewritePatternBase;
-
   LogicalResult matchAndRewrite(mesh::ProcessLinearIndexOp op,
                                 PatternRewriter &rewriter) const override {
     ImplicitLocOpBuilder builder(op->getLoc(), rewriter);
     builder.setInsertionPointAfter(op.getOperation());
-    Value channel = buildCachedChannelLoading(op, symbolTableCollection,
-                                              useNamedDefaultChannels, builder);
+    Value channel = buildCachedChannelLoading(op, builder);
     Value newIndex = builder.create<IREE::Flow::ChannelRankOp>(
         builder.getIndexType(), channel);
     rewriter.replaceAllUsesWith(op.getResult(), newIndex);
@@ -544,12 +352,10 @@ struct MeshReduceScatterToFlow
     : MeshToFlowCollectiveRewritePatternBase<mesh::ReduceScatterOp> {
   using MeshToFlowCollectiveRewritePatternBase::
       MeshToFlowCollectiveRewritePatternBase;
-
   LogicalResult matchAndRewrite(mesh::ReduceScatterOp op,
                                 PatternRewriter &rewriter) const override {
-    if (ShapedType::isDynamicShape(
-            cast<RankedTensorType>(op.getOperand().getType()).getShape()) ||
-        ShapedType::isDynamicShape(op.getResult().getType().getShape())) {
+    if (!cast<RankedTensorType>(op.getOperand().getType()).hasStaticShape() ||
+        !op.getResult().getType().hasStaticShape()) {
       // TODO: add dynamic support.
       return rewriter.notifyMatchFailure(op->getLoc(),
                                          "Dynamic tensor case is unsupported.");
@@ -557,8 +363,7 @@ struct MeshReduceScatterToFlow
 
     ImplicitLocOpBuilder builder(op->getLoc(), rewriter);
     builder.setInsertionPointAfter(op.getOperation());
-    Value channel = buildCachedChannelLoading(op, symbolTableCollection,
-                                              useNamedDefaultChannels, builder);
+    Value channel = buildCachedChannelLoading(op, builder);
 
     int64_t scatterAxis = op.getScatterAxis().getSExtValue();
 
@@ -586,157 +391,17 @@ struct MeshReduceScatterToFlow
   }
 };
 
-using MeshAndAxesSet =
-    DenseSet<std::tuple<mesh::MeshOp, SmallVector<mesh::MeshAxis>>>;
-
-template <typename Op>
-struct CollectiveOpVisitor {
-  CollectiveOpVisitor(MeshAndAxesSet &meshAndAxesSet,
-                      SymbolTableCollection &symbolTableCollection)
-      : meshAndAxesSet(meshAndAxesSet),
-        symbolTableCollection(symbolTableCollection) {}
-  void operator()(Op op) {
-    meshAndAxesSet.insert(std::make_tuple(
-        symbolTableCollection.lookupNearestSymbolFrom<mesh::MeshOp>(
-            op, op.getMeshAttr()),
-        llvm::to_vector(op.getMeshAxes())));
-  }
-
-private:
-  MeshAndAxesSet &meshAndAxesSet;
-  SymbolTableCollection &symbolTableCollection;
-};
-
-template <typename Op>
-struct CollectiveOpWithoutMeshAxesVisitor {
-  CollectiveOpWithoutMeshAxesVisitor(
-      MeshAndAxesSet &meshAndAxesSet,
-      SymbolTableCollection &symbolTableCollection)
-      : meshAndAxesSet(meshAndAxesSet),
-        symbolTableCollection(symbolTableCollection) {}
-  void operator()(Op op) {
-    mesh::MeshOp mesh =
-        symbolTableCollection.lookupNearestSymbolFrom<mesh::MeshOp>(
-            op, op.getMeshAttr());
-    meshAndAxesSet.insert(std::make_tuple(mesh, getAllMeshAxes(mesh)));
-  }
-
-private:
-  MeshAndAxesSet &meshAndAxesSet;
-  SymbolTableCollection &symbolTableCollection;
-};
-
-void populateMeshAndAxes(Operation *op, MeshAndAxesSet &meshAndAxesSet,
-                         SymbolTableCollection &symbolTableCollection) {
-  OpVisitorCollection opVisitors;
-  opVisitors.emplaceVisitors<
-      CollectiveOpVisitor<mesh::AllGatherOp>,
-      CollectiveOpVisitor<mesh::AllReduceOp>,
-      CollectiveOpVisitor<mesh::AllToAllOp>,
-      CollectiveOpVisitor<mesh::ReduceScatterOp>,
-      CollectiveOpWithoutMeshAxesVisitor<mesh::ProcessLinearIndexOp>>(
-      meshAndAxesSet, symbolTableCollection);
-
-  op->walk([&opVisitors](Operation *op) {
-    opVisitors(op);
-    return WalkResult::advance();
-  });
-}
-
-static void createChannels(ModuleOp moduleOp,
-                           SymbolTableCollection &symbolTableCollection,
-                           MeshAndAxesSet &meshAndAxesSet,
-                           bool useNamedDefaultChannels) {
-  populateMeshAndAxes(moduleOp, meshAndAxesSet, symbolTableCollection);
-
-  OpBuilder builder(moduleOp->getContext());
-
-  // Sort for deterministic testing with FileCheck.
-  auto meshAndAxesSetSorted = llvm::to_vector(meshAndAxesSet);
-  llvm::sort(meshAndAxesSetSorted, [](auto &a, auto &b) {
-    int nameCompareRes =
-        std::get<0>(a).getSymName().compare(std::get<0>(b).getSymName());
-    if (nameCompareRes == 0)
-      return std::get<1>(a) < std::get<1>(b);
-    return nameCompareRes < 0;
-  });
-  for (auto &[mesh, meshAxes] : llvm::make_range(meshAndAxesSetSorted.rbegin(),
-                                                 meshAndAxesSetSorted.rend())) {
-    buildGlobalChannelCreation(mesh, meshAxes, useNamedDefaultChannels,
-                               moduleOp, builder);
-  }
-}
-
-static LogicalResult
-convertCollectives(ModuleOp moduleOp,
-                   SymbolTableCollection &symbolTableCollection,
-                   bool useNamedDefaultChannels) {
-  RewritePatternSet patterns(moduleOp->getContext());
-  IREE::Flow::populateMeshToFlowCollectivesPatterns(
-      patterns, symbolTableCollection, useNamedDefaultChannels);
-  return applyPatternsAndFoldGreedily(moduleOp, std::move(patterns));
-}
-
-static void removeMeshOps(MeshAndAxesSet &meshAndAxesSet) {
-  auto meshRange =
-      llvm::map_range(meshAndAxesSet, [](auto &v) { return std::get<0>(v); });
-  DenseSet<mesh::MeshOp> meshOpsSet(std::begin(meshRange), std::end(meshRange));
-  for (mesh::MeshOp op : meshOpsSet) {
-    if (op)
-      op.erase();
-  }
-}
-
-struct ConvertMeshToFlowPass
-    : public IREE::Flow::ConvertMeshToFlowBase<ConvertMeshToFlowPass> {
-  void getDependentDialects(DialectRegistry &registry) const override {
-    registerMeshToFlowDependencies(registry);
-  }
-
-  void runOnOperation() override {
-    // Run only on the top module.
-    if (getOperation()->getParentOp() != nullptr) {
-      return;
-    }
-
-    MeshAndAxesSet meshAndAxesSet;
-    SymbolTableCollection symbolTableCollection;
-    bool useNamedDefaultChannels = hasMoreThanOneMesh(getOperation());
-
-    createChannels(getOperation(), symbolTableCollection, meshAndAxesSet,
-                   useNamedDefaultChannels);
-    if (failed(convertCollectives(getOperation(), symbolTableCollection,
-                                  useNamedDefaultChannels))) {
-      return signalPassFailure();
-    }
-
-    // Cleanup mesh definition ops that are no longer referenced.
-    removeMeshOps(meshAndAxesSet);
-  }
-};
-
 } // namespace
 
 void populateMeshToFlowCollectivesPatterns(
     RewritePatternSet &patterns, SymbolTableCollection &symbolTableCollection,
-    bool useNamedDefaultChannels) {
+    LookupMeshChannelFn lookupChannel) {
   patterns.add<MeshAllGatherToFlow, MeshAllReduceToFlow, MeshAllToAllToFlow,
                MeshReduceScatterToFlow, MeshProcessLinearIndexToFlow>(
-      symbolTableCollection, useNamedDefaultChannels, patterns.getContext());
+      symbolTableCollection, lookupChannel, patterns.getContext());
   mesh::populateFoldingPatterns(patterns, symbolTableCollection);
   mesh::populateProcessMultiIndexOpLoweringPatterns(patterns,
                                                     symbolTableCollection);
 }
-
-std::unique_ptr<Pass> createConvertMeshToFlowPass() {
-  return std::make_unique<ConvertMeshToFlowPass>();
-}
-
-void registerMeshToFlowDependencies(DialectRegistry &registry) {
-  registry.insert<affine::AffineDialect, FlowDialect, linalg::LinalgDialect,
-                  mesh::MeshDialect, tensor::TensorDialect>();
-}
-
-void registerMeshToFlowPasses() { registerPasses(); }
 
 } // namespace mlir::iree_compiler::IREE::Flow
