+++ compiler/src/iree/compiler/Dialect/Flow/Transforms/Passes.td
@@ -53,6 +53,53 @@ def CollapseReductionDimensionsPass :
   ];
 }
 
+def ConvertMeshToFlowPass :
+    Pass<"iree-convert-mesh-to-flow", "mlir::ModuleOp"> {
+  let summary = "Convert Mesh dialect operations to flow.";
+  let description = [{
+    Each mesh corresponds to a default !flow.channel with the same group name.
+    ```
+    mesh.mesh @mesh_1(shape = 2x3)
+    ```
+    ```
+    %channel = flow.channel.default "mesh_1" : !flow.channel
+    ```
+    If there is onl one mesh in the program than the name is omitted and the
+    ```
+    %channel = flow.channel.default : !flow.channel
+    ```
+
+    Each (mesh, mesh_axes) pair partitions and orders the devices into disjoint
+    groups, each corresponding to a !flow.channel to perform a collective
+    operation.
+    For example
+    ```
+    mesh.mesh @mesh(shape = 2x3x4x5)
+    ...
+    %1 = mesh.all_reduce on @mesh mesh_axes = [2, 0] : tensor<10x20xf32>
+    ```
+    For more information see
+    [Mesh dialect](https://mlir.llvm.org/docs/Dialects/Mesh/#device-groups).
+
+    The mesh partition and device ordering determines the values for the
+    `color` and `key` in the corresponding `flow.channel.split` operation used
+    to create the channel.
+    For more information on the meaning of `color` and `key` see
+    [MPI_Comm_split](https://www.mpi-forum.org/docs/mpi-4.1/mpi41-report/node188.htm#Node188)
+    in the MPI standard.
+
+    Each !flow.channel is wrapped in an IREE `util.global` and its construction
+    is done only once with `util.initializer`.
+  }];
+  let dependentDialects = [
+    "mlir::affine::AffineDialect",
+    "mlir::linalg::LinalgDialect",
+    "mlir::mesh::MeshDialect",
+    "mlir::tensor::TensorDialect",
+    "IREE::Flow::FlowDialect",
+  ];
+}
+
 def ConvertRegionToWorkgroupsPass :
     Pass<"iree-flow-convert-region-to-workgroups", ""> {
   let summary = "Convert dispatch.region ops to dispatch.workgroups ops.";
