+++ compiler/src/iree/compiler/Tools/init_iree_passes.h
@@ -17,7 +17,6 @@
 #include "iree/compiler/Bindings/Native/Transforms/Passes.h"
 #include "iree/compiler/Bindings/TFLite/Transforms/Passes.h"
 #include "iree/compiler/ConstEval/Passes.h"
-#include "iree/compiler/Dialect/Flow/Conversion/MeshToFlow/MeshToFlow.h"
 #include "iree/compiler/Dialect/Flow/Transforms/Passes.h"
 #include "iree/compiler/Dialect/HAL/Transforms/Passes.h"
 #include "iree/compiler/Dialect/LinalgExt/Transforms/Passes.h"
@@ -54,7 +53,6 @@ inline void registerAllIreePasses() {
   GlobalOptimization::registerGlobalOptimizationPipeline();
   Preprocessing::registerPreprocessingPasses();
   IREE::Flow::registerFlowPasses();
-  IREE::Flow::registerMeshToFlowPasses();
   IREE::HAL::registerHALPasses();
   IREE::HAL::Inline::registerHALInlinePasses();
   IREE::HAL::Loader::registerHALLoaderPasses();
